LayoutTests/editing/selection/modify-move/move-forward-after-line-break.html

<!DOCTYPE html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="../../../resources/assert-selection.js"></script>
<script>
test(() => {
  assert_selection(
    '<div contenteditable style="width:50px">|foo bar baz foo</div>',
    selection => {
      selection.modify('move', 'forward', 'line');
      selection.modify('move', 'forward', 'character');
    },
    '<div contenteditable style="width:50px">foo bar b|az foo</div>');
}, 'Modify move forward after line break');
</script>
